🧩 Modularized and refactored the hooks and utils directories to establish a cleaner, more coherent file structure and improve overall project organization.
