+++ src/components/pages/tokenomics/RevenueStreams.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState, useMemo } from 'react';
 import { motion as m } from 'framer-motion';
 import { Pie } from 'react-chartjs-2';
-import { chartOptions, revenueStreamsData } from '../../../utils/chartConfig';
-import '../../../utils/chartSetup';
+import { chartOptions, revenueStreamsData } from '../../../utils/chart/chartConfig';
+import '../../../utils/chart/chartSetup';
 
 const RevenueStreams = () => {
   const chartRef = useRef(null);
