+++ src/utils/errors/eventHandlers.js
